requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.4
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.0.0
psycopg2-binary==2.9.10
reportlab==4.2.5
sqlparse==0.5.3
tzdata==2024.2
